Extract stats fetching for ribbon from main file
